ProyectoColas/proyectoColas/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mssql-django==1.5
PyJWT==2.8.0
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
